[FMT][FIX] Fixes offest formatting to remove compilation warnings

+++ readj.c
@@ -90,7 +90,7 @@ void readJlambda(int nspect, double *J)
   if (!result) {
     sprintf(messageStr,
 	    "Error reading file: offset = %lld, recordsize = %zu",
-	    offset, recordsize);
+	    (long long)offset, recordsize);
     Error(ERROR_LEVEL_2, routineName, messageStr);
   }
 }
@@ -114,7 +114,7 @@ void writeJlambda(int nspect, double *J)
   if (!result) {
     sprintf(messageStr,
 	    "Error writing file: offset = %lld, recordsize = %zu",
-	    offset, recordsize);
+	    (long long)offset, recordsize);
     Error(ERROR_LEVEL_2, routineName, messageStr);
   }
 }
@@ -177,7 +177,7 @@ void readJ20lambda(int nspect, double *J20)
   if (!result) {
     sprintf(messageStr,
 	    "Error reading file: offset = %lld, recordsize = %zu",
-	    offset, recordsize);
+	    (long long)offset, recordsize);
     Error(ERROR_LEVEL_2, routineName, messageStr);
   }
 }
@@ -202,7 +202,7 @@ void writeJ20lambda(int nspect, double *J20)
   if (!result) {
     sprintf(messageStr,
 	    "Error writing file: offset = %lld, recordsize = %zu",
-	    offset, recordsize);
+	    (long long)offset, recordsize);
     Error(ERROR_LEVEL_2, routineName, messageStr);
   }
 }
@@ -231,7 +231,7 @@ void readImu(int nspect, int mu, bool_t to_obs, double *I)
  if (!result) {
     sprintf(messageStr,
 	    "Error reading file: offset = %lld, recordsize = %zu",
-	    offset, recordsize);
+	    (long long)offset, recordsize);
     Error(ERROR_LEVEL_2, routineName, messageStr);
   }
 }
@@ -260,7 +260,7 @@ void writeImu(int nspect, int mu, bool_t to_obs, double *I)
  if (!result) {
     sprintf(messageStr,
 	    "Error writing file: offset = %lld, recordsize = %zu",
-	    offset, recordsize);
+	    (long long)offset, recordsize);
     Error(ERROR_LEVEL_2, routineName, messageStr);
   }
 }
